BLE: Move ble::Duration in its own file.
